requirements.txt
gym==0.21.0
pytorch_lightning==1.6.1
wandb===0.13.10
omegaconf==2.3.0
fire==0.5.0
hydra-core==1.3.2
torchvision==0.15.2
einops
array2gif
minigrid 
pydub
matplotlib
names
seaborn
joblib
